🚚[Rename] api 코드 분리
