+++ src/features/club-create/hooks/useCreateBookClub.ts
@@ -1,9 +1,8 @@
 'use client';
 
 import { useState } from 'react';
-import { toKoreanTime } from '@/lib/utils/dateUtils';
 import { BookClubForm } from '../types';
-import apiClient from '@/lib/utils/apiClient';
+import { createBookClub } from '../api';
 
 export const useCreateBookClub = () => {
   const [isLoading, setIsLoading] = useState(false);
@@ -14,42 +13,9 @@ export const useCreateBookClub = () => {
     setError(null);
 
     try {
-      const formData = new FormData();
-
-      const imageFile = data.image instanceof File ? data.image : null;
-      if (imageFile) {
-        formData.append('image', imageFile);
-      }
-
-      const bookClubData = {
-        title: data.title,
-        description: data.description,
-        bookClubType: data.bookClubType,
-        meetingType: data.meetingType,
-        targetDate: toKoreanTime(data.targetDate),
-        endDate: toKoreanTime(data.endDate),
-        memberLimit: data.memberLimit,
-        city: data.city,
-        town: data.town,
-      };
-
-      formData.append(
-        'bookClub',
-        new Blob([JSON.stringify(bookClubData)], {
-          type: 'application/json',
-        }),
-      );
-
-      const response = await apiClient.post('/book-clubs', formData, {
-        headers: {
-          'Content-Type': 'multipart/form-data',
-          Authorization:
-            'Bearer eyJhbGciOiJIUzI1NiJ9.eyJ0eXBlIjoiQUNDRVNTIiwic3ViIjoiY2xvdWQwNDA2QG5hdmVyLmNvbSIsImlhdCI6MTczNDI0NzY3OSwiZXhwIjoxNzM0MzM0MDc5fQ.LPfdAD6zrsU3tOKys-WSCt6fi1QhtpVJd2VTXKYrNQc',
-        },
-      });
-
-      // TODO:: 성공 시 처리
-      return response.data;
+      const response = await createBookClub(data);
+      alert('북클럽이 성공적으로 생성되었습니다!');
+      return response;
     } catch (error) {
       setError('북클럽 생성에 실패했습니다.');
       throw error;
